+++ app/layout.tsx
@@ -14,7 +14,7 @@ const geistMono = localFont({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "Countries.fyi",
   description: "Generated by create next app",
 };
